Add autoplay function and useful animator events
The presentation and its animations will start automatically
when the user presses "A".
When "A" is pressed again, or when any deck default action key
is pressed during the autoplay, it will stop.
Some events are added to the Animator class in order to
enable this functionality in the deck animator plugin.