Create a stub of animator
Contrary to the asvg plugin upon which it is based,
the animator plugin can animate HTML elements instead
of SVG elements of an object tag.
It is otherwise used in much the same way. The name
of the javascript variable containing the Animator
object must be the value of a custom HTML parameter,
data-dahu-animator, of the slide in which to animate
elements.