Fix animation autoplay for slides without an animator
The autoplay used to stop working when several consecutive slides
had no animator / an empty animator. This is fixed.

This fix also corrects and refactors a common test used for animators:
an empty animator should be treated the same as an empty animator
in most cases. Otherwise, it might perturb the transition between
slides in the case where the presentation doesn't start at the
first slide (by means of an anchor in the URL).